Refactor the GitHub STL into multiple modules for readability.
